商城第一版综合

+++ learn-modules/learn-authorization/learn-authorization-api/src/main/java/me/own/learn/authorization/controller/AdminLoginController.java
@@ -1,6 +1,7 @@
 package me.own.learn.authorization.controller;
 
 import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiOperation;
 import me.own.learn.authorization.dto.AdminLoginDto;
 import me.own.learn.authorization.service.AdminAuthenticationRequired;
@@ -9,6 +10,7 @@ import me.own.learn.authorization.service.TokenService;
 import me.own.learn.authorization.service.model.AdminAccessToken;
 import me.own.learn.authorization.utils.CookieUtils;
 import me.own.learn.authorization.vo.AdminTokenVo;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,6 +21,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * @author yexudong
@@ -43,21 +47,28 @@ public class AdminLoginController {
             @RequestBody AdminLoginDto loginBo
     ) {
         AdminTokenVo token = adminLoginService.login(loginBo);
-        CookieUtils.setAdminUserTokenInCookie(response, token.getValue());
+        CookieUtils.setAdminUserTokenInCookie(request, response, token.getValue());
         return new ResponseEntity(token, HttpStatus.CREATED);
     }
 
     @ApiOperation("管理员登出")
     @RequestMapping(method = RequestMethod.DELETE)
     @AdminAuthenticationRequired
+//    @ApiImplicitParam(name = "a_id", value = "调试模式", paramType = "query", dataType = "String", defaultValue = "1")
     public ResponseEntity logout(
             HttpServletRequest request,
             HttpServletResponse response,
             AdminAccessToken aat
     ) {
+        Map<String, Object> responseBody = new HashMap<>();
+        if (StringUtils.isEmpty(aat.getValue())) {
+            responseBody.put("success", false);
+            responseBody.put("msg", "异常退出");
+            return new ResponseEntity(responseBody, HttpStatus.BAD_REQUEST);
+        }
         tokenService.remove(aat.getValue());
         // set empty string to cookie(delete the token cookie)
-        CookieUtils.setAdminUserTokenInCookie(response, "");
+        CookieUtils.setAdminUserTokenInCookie(request, response, "");
         return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
 }
